+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added option to connect your existing NIP-05 username.
 - Added a new UI for replying to messages that allows attaching images and setting an expiration date.
 
 ## [0.1.7] - 2024-03-21Z